# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  HTTP400BadRequest:
    properties:
      err_code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
...
  HTTP500InternalServerError:
    properties:
      err_code:
        example: 500
        type: integer
      message:
        example: InternalServerError
        type: string
    type: object
...
  ablecloud_io_ablestack-api_internal_model_cube.CCVMSnapRequest:
    properties:
      action:
        description: 'action: list/backup/rollback'
        example: backup
        type: string
      snap_name:
        description: snapshot name for rollback, or optional custom name for backup
        example: 2026-04-29-10:00:00
        type: string
    type: object
  ablecloud_io_ablestack-api_internal_model_cube.CCVMSnapResponse:
    properties:
      action:
        description: 수행한 action
        example: backup
        type: string
      code:
        description: 처리 결과 코드
        example: 200
        type: integer
      deleted_snapshots:
        description: backup 후 보관 정책으로 삭제한 snapshot 이름 목록
        items:
          type: string
        type: array
      message:
        description: 처리 결과 메시지
        example: CCVM Snapshot Backup Create Success
        type: string
      snap_name:
        description: 생성 또는 복구 대상 snapshot 이름
        example: 2026-04-29-10:00:00
        type: string
      target:
        description: 실제 작업을 수행한 대상 ablecube IP 또는 식별자
        example: 10.10.31.1
        type: string
      val:
        description: 기존 Python createReturn 호환 결과값
    type: object
...
info:
  contact:
    email: ycyun@ablecloud.io
    name: API Support
    url: https://www.ablecloud.io/support
  description: This is a Cube-API server.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://ablecloud.io/
  title: Cube API
  version: "1.0"
paths:
...
  /cube/ccvm/snap:
    post:
      consumes:
      - application/json
      description: HCI/HCI-filesystem 환경에서 CCVM RBD snapshot list/backup/rollback
        작업을 수행합니다.
      parameters:
      - description: ccvm snapshot request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.CCVMSnapRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.CCVMSnapResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: CCVM Snapshot
      tags:
      - CUBE - CCVM
...
schemes:
- http
- https
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    description: Enter the token with the Bearer prefix, e.g. Bearer eyJ...
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"